requirements.txt
slack_sdk>=3.0
slack_bolt>=1.6.1
certifi~=2022.6.15.1
python-dotenv~=0.20.0